requirements.txt
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
# azure-cli-core
azure-cli==2.36.0